Release 2.4.8 Fix problem for BLIK code starting with 0

+++ src/includes/gateways/class-wc-payment-gateway-pay-by-paynow-pl-blik.php
@@ -32,7 +32,7 @@ class WC_Payment_Gateway_Pay_By_Paynow_PL_Blik extends WC_Gateway_Pay_By_Paynow_
 	 * @return bool
 	 */
 	public function validate_fields(): bool {
-		$payment_authorization_code = (int) preg_replace( '/\s+/', '', filter_input( INPUT_POST, 'authorizationCode' ) );
+		$payment_authorization_code = preg_replace( '/\s+/', '', filter_input( INPUT_POST, 'authorizationCode' ) );
 		$blik_payment_methods       = $this->get_only_payment_methods_for_type( Type::BLIK );
 		if ( $blik_payment_methods && $this->isWhiteLabelEnabled( $blik_payment_methods ) &&
 			( empty( $payment_authorization_code ) || strlen( $payment_authorization_code ) !== 6 ) ) {
